Move FulfillmentErrorCode to rustc_trait_selection too
